Refactor CartItemCard styles
