--- components/CartItemCard/CartItemCard.styled.js
@@ -1,59 +1,0 @@
-// Import external resources
-import Image from "next/image";
-// Import internal resources
-import styled from "styled-components";
-import Link from "next/link";
-
-export const Thumbnail = styled(Image)`
-  border-radius: 8px;
-  object-fit: cover;
-`;
-
-export const CartItemArticle = styled.article`
-  border: solid lightgray;
-  border-radius: 16px;
-  width: 80vw;
-  padding: 10px;
-
-  display: grid;
-  grid-template-columns: 50% 50%;
-  place-items: center;
-`;
-
-export const CartItemLeft = styled.div`
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
-`;
-
-export const CartItemRight = styled.div`
-  height: 100%;
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding-top: 10px;
-`;
-
-export const CartItemParagraph = styled.p`
-  margin: 0;
-  text-align: center;
-`;
-
-export const CartItemDeleteButton = styled.button`
-  align-self: flex-end;
-`;
-
-export const ShoppingCartLink = styled(Link)`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
-  color: inherit;
-  text-decoration: inherit;
-`;
